requirements.txt
# Core dependencies
numpy>=1.22.0
scipy>=1.8.0
librosa>=0.9.0
soundfile>=0.10.0

# Audio processing
pyaudio>=0.2.11
wave>=0.0.2

# Speaker diarization
pyannote.audio>=3.1.0
torch>=1.12.0
torchaudio>=0.12.0

# Voice embeddings
resemblyzer>=0.1.1

# Speech transcription
faster-whisper>=0.9.0
ctranslate2>=3.0.0

# Database
sqlite3

# Utilities
tqdm>=4.64.0
requests>=2.28.0 